--- src/pages/Search/ScrollButton.tsx
@@ -1,30 +1,0 @@
-import theme from "tools/theme";
-
-import ArrowUpwardRoundedIcon from "@mui/icons-material/ArrowUpwardRounded";
-
-const ScrollButton = () => {
-  return (
-    <div
-      style={{
-        position: "fixed",
-        bottom: 72,
-        right: 16,
-        zIndex: theme.zIndex_nav,
-        height: 48,
-        width: 48,
-        borderRadius: "50%",
-        backgroundColor: theme.white,
-        boxShadow: theme.shadow_clicked,
-        display: "flex",
-        justifyContent: "center",
-        alignItems: "center",
-        ...theme.cursor(),
-      }}
-      onClick={() => window.scrollTo({ top: 0, behavior: "smooth" })}
-    >
-      <ArrowUpwardRoundedIcon style={{ color: theme.purple, fontSize: 28 }} />
-    </div>
-  );
-};
-
-export default ScrollButton;
