+++ src/pages/Search/SideResult.jsx
@@ -5,12 +5,11 @@ import usePageFromSearchParams from "hooks/usePageFromSearchParams";
 
 import DottedLine from "components/DottedLine";
 import Empty from "components/Empty";
+import ModalRoomSelection from "components/ModalPopup/ModalRoomSelect";
 import Pagination, { PAGE_MAX_ITEMS } from "components/Pagination";
 import Room from "components/Room";
 import Title from "components/Title";
 import WhiteContainer from "components/WhiteContainer";
-
-import RoomSelectionModal from "./RoomSelectionModal";
 
 import theme from "tools/theme";
 
@@ -163,7 +162,7 @@ const SideResult = (props) => {
   if (!props.mobile) {
     return (
       <div style={{ marginTop: 26 }}>
-        <RoomSelectionModal
+        <ModalRoomSelection
           isOpen={!!selectedRoomInfo}
           onClose={() => {
             setSelectedRoomInfo(null);
@@ -210,7 +209,7 @@ const SideResult = (props) => {
   } else {
     return (
       <>
-        <RoomSelectionModal
+        <ModalRoomSelection
           isOpen={!!selectedRoomInfo}
           onClose={() => {
             setSelectedRoomInfo(null);
