+++ src/pages/Search/index.jsx
@@ -1,6 +1,5 @@
-import PropTypes from "prop-types";
 import qs from "qs";
-import { memo, useEffect, useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import { useCookies } from "react-cookie";
 import { useHistory, useLocation } from "react-router-dom";
 
@@ -16,131 +15,19 @@ import {
   OptionTime,
 } from "components/ModalRoomOptions";
 import RLayout from "components/RLayout";
+import ScrollUpButton from "components/ScrollUpButton";
 import Title from "components/Title";
 import Tooltip from "components/Tooltip";
 
-import ScrollButton from "./ScrollButton";
+import SelectSearchOptions from "./SelectSearchOptions";
 import SideResult from "./SideResult";
+import { isSearchAll, isValidQuery } from "./utils";
 
-import hoverEventSet from "tools/hoverEventSet";
 import moment, { getToday, getToday10 } from "tools/moment";
 import theme from "tools/theme";
 
 const searchQueryOption = { skipNulls: true };
 const defaultOptions = { place: true, date: true, time: true };
-
-const SearchOption = (props) => {
-  const [isHover, setHover] = useState(false);
-  const style = {
-    ...theme.font12,
-    borderRadius: "15px",
-    padding: "8px 15px 7px 15px",
-    boxShadow: theme.shadow,
-    background: props.selected
-      ? isHover
-        ? theme.purple_dark
-        : theme.purple
-      : isHover
-      ? theme.purple_hover
-      : theme.white,
-    color: props.selected ? theme.white : theme.black,
-    config: { duration: 150 },
-    ...theme.cursor(),
-  };
-  const onClick = () => {
-    props.handler((prevState) => {
-      const _options = { ...prevState };
-      _options[props.id] = !props.selected;
-      if (!_options.date && _options.time) {
-        if (props.id === "date") _options.time = false;
-        if (props.id === "time") _options.date = true;
-      }
-      return _options;
-    });
-  };
-  return (
-    <div style={style} onClick={onClick} {...hoverEventSet(setHover)}>
-      {props.children}
-    </div>
-  );
-};
-SearchOption.propTypes = {
-  children: PropTypes.string,
-  id: PropTypes.string,
-  selected: PropTypes.bool,
-  handler: PropTypes.func,
-};
-const MemoizedSearchOption = memo(SearchOption);
-
-const SelectSearchOptions = (props) => {
-  const options = [
-    { name: "장소", id: "place" },
-    { name: "날짜", id: "date" },
-    { name: "시간", id: "time" },
-    { name: "최대 인원", id: "maxPeople" },
-    { name: "방 이름", id: "name" },
-  ];
-  return (
-    <div
-      style={{
-        display: "flex",
-        flexWrap: "wrap",
-        gap: "10px",
-        paddingTop: "10px",
-        paddingBottom: "15px",
-      }}
-    >
-      {options.map((item, index) => {
-        const selected = props.options[item.id] ?? false;
-        return (
-          <MemoizedSearchOption
-            key={index}
-            id={item.id}
-            handler={props.handler}
-            selected={selected}
-          >
-            {item.name}
-          </MemoizedSearchOption>
-        );
-      })}
-    </div>
-  );
-};
-SelectSearchOptions.propTypes = {
-  options: PropTypes.object,
-  handler: PropTypes.func,
-};
-
-const isSearchAll = (q) => {
-  for (let [key, val] of Object.entries(q))
-    if (key === "all" && val === "true") return true;
-  return false;
-};
-
-const isValidQuery = (q) => {
-  const allowedKeys = [
-    "name",
-    "from",
-    "to",
-    "time",
-    "withTime",
-    "maxPeople",
-    "page",
-  ];
-  const keys = Object.keys(q);
-
-  if (keys.length > allowedKeys.length) return false;
-  if (keys.some((key) => !allowedKeys.includes(key))) return false;
-  if (keys.includes("from") !== keys.includes("to")) return false;
-  if (keys.includes("maxPeople") && q.maxPeople !== null) {
-    const parsedInt = parseInt(q.maxPeople);
-    if (isNaN(parsedInt) || parsedInt < 2 || parsedInt > 4) return false;
-  }
-  if (keys.includes("time") && q.time !== null) {
-    if (isNaN(Date.parse(q.time))) return false;
-  } else if (keys.includes("withTime") && q.withTime === "true") return false;
-  return true;
-};
 
 const Search = () => {
   const reactiveState = useR2state();
@@ -266,7 +153,7 @@ const Search = () => {
     } else {
       history.replace("/search");
     }
-  }, [JSON.stringify(location)]);
+  }, [location.search]);
 
   useEffect(() => {
     if (!Object.values(searchOptions).some((option) => option)) {
@@ -422,17 +309,13 @@ const Search = () => {
         {message}
       </Button>
       {!Object.values(searchOptions).some((option) => option) && (
-        <Tooltip
-          text={
-            "검색 옵션을 선택하지 않을 경우 '빠른 출발 검색'이 가능합니다. 현재 시각에서 24시간 내의 방들이 검색됩니다."
-          }
-        />
+        <Tooltip text="검색 옵션을 선택하지 않을 경우 '빠른 출발 검색'이 가능합니다. 현재 시각에서 24시간 내의 방들이 검색됩니다." />
       )}
       {searchResult && reactiveState === 3 && (
         <div style={{ paddingTop: "30px" }} ref={scrollRef}>
           <Title icon="search_result">검색 결과</Title>
           <SideResult result={searchResult} mobile />
-          {showScrollButton && <ScrollButton />}
+          {showScrollButton && <ScrollUpButton />}
         </div>
       )}
     </>
@@ -441,12 +324,12 @@ const Search = () => {
     <SideResult result={searchResult} />
   );
   return (
-    <div>
+    <>
       <Title icon="search" header marginAuto R2={searchResult !== null}>
         방 검색하기
       </Title>
       <RLayout.R2 left={leftLay} right={rightLay} />
-    </div>
+    </>
   );
 };
